Prepare alloc library for paging system
