+++ link.ld
@@ -1,4 +1,4 @@
-PROVIDE(__heap_size = 2 * 1024 * 1024);
+PROVIDE(__pagespace_size = 512 * 1024 * 1024);
 SECTIONS
 {
     . = 0x80000;     /* Kernel load address for AArch64 */
@@ -14,12 +14,11 @@ SECTIONS
         *(COMMON)
         __bss_end = .;
     }
-    .heap (NOLOAD) : {
+    .pages (NOLOAD) : {
         . = ALIGN(16);
-        PROVIDE(__heap_start = .);
-        *(.heap)
-        . = __heap_start + __heap_size;
-        __heap_end = .;
+        PROVIDE(__pagespace_start = .);
+        . = __pagespace_start + __pagespace_size;
+        __pagespace_end = .;
     }
     _end = .;
